Moved project details to separate file
