+++ mochi_code/code/__init__.py
@@ -1,0 +1,4 @@
+"""__init__.py"""
+from .project_details import ProjectDetails
+
+__all__ = ["ProjectDetails"]
